Upgraded coingecko to v3 framework (#2631)
