+++ .changeset/shiny-queens-return.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/coingecko-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
