+++ packages/core/legos/src/sources.ts
@@ -36,7 +36,6 @@ import * as coinapi_test from '@chainlink/coinapi-test-adapter'
 import * as coinbase from '@chainlink/coinbase-adapter'
 import * as coincodex from '@chainlink/coincodex-adapter'
 import * as coingecko from '@chainlink/coingecko-adapter'
-import * as coingecko_test from '@chainlink/coingecko-test-adapter'
 import * as coinlore from '@chainlink/coinlore-adapter'
 import * as coinmarketcap from '@chainlink/coinmarketcap-adapter'
 import * as coinmetrics from '@chainlink/coinmetrics-adapter'
@@ -203,7 +202,6 @@ export default {
   coinbase,
   coincodex,
   coingecko,
-  coingecko_test,
   coinlore,
   coinmarketcap,
   coinmetrics,
