+++ packages/core/legos/tsconfig.json
@@ -125,9 +125,6 @@
       "path": "../../sources/coingecko"
     },
     {
-      "path": "../../sources/coingecko-test"
-    },
-    {
       "path": "../../sources/coinlore"
     },
     {
