--- packages/sources/coingecko-test/CHANGELOG.md
@@ -1,155 +1,0 @@
-# @chainlink/coingecko-test-adapter
-
-## 1.6.6
-
-### Patch Changes
-
-- 4245b4d8b: Bumped framework version
-
-## 1.6.5
-
-### Patch Changes
-
-- 8fe5be225: Bumped framework version
-
-## 1.6.4
-
-### Patch Changes
-
-- c9cee149e: Bump fw version
-
-## 1.6.3
-
-### Patch Changes
-
-- 2fdaa5aa4: Bump v3 version
-
-## 1.6.2
-
-### Patch Changes
-
-- bc8886c63: Bumped framework version
-
-## 1.6.1
-
-### Patch Changes
-
-- 37c0d8883: Bumped v3 framework version
-
-## 1.6.0
-
-### Minor Changes
-
-- d474a5553: Update v3 framework to latest version
-
-## 1.5.4
-
-### Patch Changes
-
-- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
-
-## 1.5.3
-
-### Patch Changes
-
-- ae1cf06fb: Update EA framework version
-
-## 1.5.2
-
-### Patch Changes
-
-- e018cf5b4: Fix override ID for "COMP"
-
-## 1.5.1
-
-### Patch Changes
-
-- 43f8d6d86: Fix override for "ARPA"
-
-## 1.5.0
-
-### Minor Changes
-
-- 4d02034ae: Enable configuring the API_ENDPOINT
-  Remove "precision" input param (now hard-coded as "full")
-
-## 1.4.0
-
-### Minor Changes
-
-- 045897b8e: Include "result" in "data" object
-  Include "precision" in data provider request
-
-## 1.3.2
-
-### Patch Changes
-
-- b98296089: Bumped v3 framework version
-
-## 1.3.1
-
-### Patch Changes
-
-- 255af9e8a: Bumped v3 framework version
-- 1c73ae21f: Only cache data related to the specified request
-
-## 1.3.0
-
-### Minor Changes
-
-- dc347f5bf: Updated EA framework version
-  Cache errored responses
-  Fix FLOKI override ID
-
-### Patch Changes
-
-- 3344df573: Added more overrides and fixed overrides helper
-
-## 1.2.3
-
-### Patch Changes
-
-- 7604a7f3c: Added/fixed static overrides.
-  Ensure the coin ID is always lower-case.
-
-## 1.2.2
-
-### Patch Changes
-
-- 07781d25d: Bumped framework version
-
-## 1.2.1
-
-### Patch Changes
-
-- 4972b1fd3: Bump framework version
-
-## 1.2.0
-
-### Minor Changes
-
-- 0ffe3a98a: Bump fw version
-
-## 1.1.2
-
-### Patch Changes
-
-- d7df8c3d0: Update overrides
-
-## 1.1.1
-
-### Patch Changes
-
-- 04a8c47e2: bump fw version
-
-## 1.1.0
-
-### Minor Changes
-
-- 3150b1d08: Bump framework
-
-## 1.0.0
-
-### Major Changes
-
-- 014d53709: Initial version of the adapter
