--- packages/sources/coingecko-test/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@chainlink/coingecko-test-adapter",
-  "version": "1.6.6",
-  "description": "Chainlink coingecko-test adapter.",
-  "keywords": [
-    "Chainlink",
-    "LINK",
-    "blockchain",
-    "oracle",
-    "coingecko-test"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "url": "https://github.com/smartcontractkit/external-adapters-js",
-    "type": "git"
-  },
-  "license": "MIT",
-  "scripts": {
-    "clean": "rm -rf dist && rm -f tsconfig.tsbuildinfo",
-    "prepack": "yarn build",
-    "build": "tsc -b",
-    "server": "node -e 'require(\"./index.js\").server()'",
-    "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
-    "start": "yarn server:dist",
-    "dev": "NODE_ENV=develop tsnd --respawn --transpile-only --project tsconfig.json './src/index.ts'",
-    "test": "jest"
-  },
-  "dependencies": {
-    "@chainlink/external-adapter-framework": "0.28.2",
-    "tslib": "2.4.0"
-  },
-  "devDependencies": {
-    "@types/jest": "27.5.2",
-    "@types/node": "16.11.51",
-    "@types/supertest": "2.0.12",
-    "axios": "^1.2.0",
-    "nock": "13.2.9",
-    "supertest": "6.2.4",
-    "typescript": "4.9.5"
-  }
-}
