--- packages/sources/coingecko-test/src/endpoint/index.ts
@@ -1,6 +1,0 @@
-export { endpoint as coins } from './coins'
-export { endpoint as crypto } from './crypto'
-export { endpoint as cryptoMarketcap } from './crypto-marketcap'
-export { endpoint as globalMarketcap } from './global-marketcap'
-export { endpoint as dominance } from './dominance'
-export { endpoint as cryptoVolume } from './crypto-volume'
