--- packages/sources/coingecko-test/test/integration/adapter.test.ts
@@ -1,207 +1,0 @@
-import { mockCryptoSuccess, mockDominanceSuccess } from './fixtures'
-import { SuperTest, Test } from 'supertest'
-import { setupExternalAdapterTest, SuiteContext } from './setup'
-import { ServerInstance } from '@chainlink/external-adapter-framework'
-
-describe('execute', () => {
-  let spy: jest.SpyInstance
-  beforeAll(async () => {
-    const mockDate = new Date('2022-05-10T16:09:27.193Z')
-    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
-  })
-
-  afterAll((done) => {
-    spy.mockRestore()
-    done()
-  })
-
-  const id = '1'
-
-  const context: SuiteContext = {
-    req: null,
-    server: async () => {
-      process.env['RATE_LIMIT_CAPACITY_SECOND'] = '5000'
-      process.env['CACHE_POLLING_MAX_RETRIES'] = '20'
-      process.env['CACHE_POLLING_SLEEP_MS'] = '500'
-      process.env['METRICS_ENABLED'] = 'false'
-      process.env['BACKGROUND_EXECUTE_MS_HTTP'] = '50'
-      const server = (await import('../../src')).server
-      return server() as Promise<ServerInstance>
-    },
-  }
-
-  const envVariables = {
-    CACHE_ENABLED: 'false',
-  }
-
-  setupExternalAdapterTest(envVariables, context)
-
-  describe('crypto api', () => {
-    const data = {
-      id,
-      data: {
-        base: 'ETH',
-        quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
-      mockCryptoSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-
-    const dataWithOverride = {
-      id,
-      data: {
-        base: 'OHMV2',
-        quote: 'USD',
-        overrides: {
-          coingecko: {
-            OHMV2: 'olympus',
-          },
-        },
-      },
-    }
-
-    it('should return success for override', async () => {
-      mockCryptoSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(dataWithOverride)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-
-    const dataWithArray = {
-      id,
-      data: {
-        base: ['OHMV2', 'ETH'],
-        quote: 'USD',
-        overrides: {
-          coingecko: {
-            OHMV2: 'olympus',
-          },
-        },
-      },
-    }
-
-    it('should return failure for array', async () => {
-      mockCryptoSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(dataWithArray)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(400)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('volume api', () => {
-    const data = {
-      id,
-      data: {
-        endpoint: 'volume',
-        base: 'ETH',
-        quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
-      mockCryptoSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('marketcap api', () => {
-    const data = {
-      id,
-      data: {
-        endpoint: 'marketcap',
-        base: 'ETH',
-        quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
-      mockCryptoSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('globalmarketcap api', () => {
-    const data = {
-      id,
-      data: {
-        endpoint: 'globalmarketcap',
-        quote: 'USD',
-      },
-    }
-
-    it('should return success', async () => {
-      mockDominanceSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-
-  describe('dominance api', () => {
-    const data = {
-      id,
-      data: {
-        endpoint: 'dominance',
-        quote: 'ETH',
-      },
-    }
-
-    it('should return success', async () => {
-      mockDominanceSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
-    })
-  })
-})
