--- packages/sources/coingecko/schemas/env.json
@@ -1,17 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/coingecko-adapter.json",
-  "title": "Chainlink External Adapter for CoinGecko",
-  "required": [],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string",
-      "description": "An optional API key for the Pro API"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
