--- packages/sources/coingecko/src/config/limits.json
@@ -1,22 +1,0 @@
-{
-  "http": {
-    "free": {
-      "rateLimit1s": 10,
-      "rateLimit1m": 50,
-      "note": "1s found in ToS, 1m found at https://www.coingecko.com/en/api"
-    },
-    "analyst": {
-      "rateLimit1m": 500,
-      "rateLimit1h": 690
-    },
-    "chainlink": {
-      "rateLimit1m": 500,
-      "rateLimit1h": 4166
-    },
-    "pro": {
-      "rateLimit1m": 500,
-      "rateLimit1h": 6900
-    }
-  },
-  "ws": {}
-}
