+++ packages/sources/coingecko/src/endpoint/global-marketcap.ts
