+++ packages/sources/coingecko/src/endpoint/index.ts
@@ -1,9 +1,6 @@
-import type { TInputParameters as CoinsInputParameters } from './coins'
-import type { TInputParameters as CryptoInputParameters } from './crypto'
-import type { TInputParameters as GlobalInputParameters } from './global'
-
-export type TInputParameters = CoinsInputParameters | CryptoInputParameters | GlobalInputParameters
-
-export * as crypto from './crypto'
-export * as global from './global'
-export * as coins from './coins'
+export { endpoint as coins } from './coins'
+export { endpoint as crypto } from './crypto'
+export { endpoint as cryptoMarketcap } from './crypto-marketcap'
+export { endpoint as globalMarketcap } from './global-marketcap'
+export { endpoint as dominance } from './dominance'
+export { endpoint as cryptoVolume } from './crypto-volume'
