+++ packages/sources/coingecko/src/global-utils.ts
