+++ packages/sources/coingecko/test/integration/adapter.test.ts
@@ -1,16 +1,33 @@
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import { server as startServer } from '../../src'
 import { mockCryptoSuccess, mockDominanceSuccess } from './fixtures'
-import { setupExternalAdapterTest } from '@chainlink/ea-test-helpers'
-import type { SuiteContext } from '@chainlink/ea-test-helpers'
 import { SuperTest, Test } from 'supertest'
+import { setupExternalAdapterTest, SuiteContext } from './setup'
+import { ServerInstance } from '@chainlink/external-adapter-framework'
 
 describe('execute', () => {
+  let spy: jest.SpyInstance
+  beforeAll(async () => {
+    const mockDate = new Date('2022-05-10T16:09:27.193Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
+  })
+
+  afterAll((done) => {
+    spy.mockRestore()
+    done()
+  })
+
   const id = '1'
 
   const context: SuiteContext = {
     req: null,
-    server: startServer,
+    server: async () => {
+      process.env['RATE_LIMIT_CAPACITY_SECOND'] = '5000'
+      process.env['CACHE_POLLING_MAX_RETRIES'] = '20'
+      process.env['CACHE_POLLING_SLEEP_MS'] = '500'
+      process.env['METRICS_ENABLED'] = 'false'
+      process.env['BACKGROUND_EXECUTE_MS_HTTP'] = '50'
+      const server = (await import('../../src')).server
+      return server() as Promise<ServerInstance>
+    },
   }
 
   const envVariables = {
@@ -20,7 +37,7 @@ describe('execute', () => {
   setupExternalAdapterTest(envVariables, context)
 
   describe('crypto api', () => {
-    const data: AdapterRequest = {
+    const data = {
       id,
       data: {
         base: 'ETH',
@@ -41,14 +58,14 @@ describe('execute', () => {
       expect(response.body).toMatchSnapshot()
     })
 
-    const dataWithOverride: AdapterRequest = {
+    const dataWithOverride = {
       id,
       data: {
-        base: 'OHM',
+        base: 'OHMV2',
         quote: 'USD',
         overrides: {
           coingecko: {
-            OHM: 'olympus',
+            OHMV2: 'olympus',
           },
         },
       },
@@ -67,20 +84,20 @@ describe('execute', () => {
       expect(response.body).toMatchSnapshot()
     })
 
-    const dataWithArray: AdapterRequest = {
+    const dataWithArray = {
       id,
       data: {
-        base: ['OHM', 'ETH'],
+        base: ['OHMV2', 'ETH'],
         quote: 'USD',
         overrides: {
           coingecko: {
-            OHM: 'olympus',
+            OHMV2: 'olympus',
           },
         },
       },
     }
 
-    it('should return success for array', async () => {
+    it('should return failure for array', async () => {
       mockCryptoSuccess()
 
       const response = await (context.req as SuperTest<Test>)
@@ -89,13 +106,13 @@ describe('execute', () => {
         .set('Accept', '*/*')
         .set('Content-Type', 'application/json')
         .expect('Content-Type', /json/)
-        .expect(200)
+        .expect(400)
       expect(response.body).toMatchSnapshot()
     })
   })
 
   describe('volume api', () => {
-    const data: AdapterRequest = {
+    const data = {
       id,
       data: {
         endpoint: 'volume',
@@ -119,7 +136,7 @@ describe('execute', () => {
   })
 
   describe('marketcap api', () => {
-    const data: AdapterRequest = {
+    const data = {
       id,
       data: {
         endpoint: 'marketcap',
@@ -143,7 +160,7 @@ describe('execute', () => {
   })
 
   describe('globalmarketcap api', () => {
-    const data: AdapterRequest = {
+    const data = {
       id,
       data: {
         endpoint: 'globalmarketcap',
@@ -166,7 +183,7 @@ describe('execute', () => {
   })
 
   describe('dominance api', () => {
-    const data: AdapterRequest = {
+    const data = {
       id,
       data: {
         endpoint: 'dominance',
