--- packages/sources/coingecko/test/unit/global.test.ts
@@ -1,36 +1,0 @@
-import { AdapterError, Requester } from '@chainlink/ea-bootstrap'
-import { assertError } from '@chainlink/ea-test-helpers'
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import { makeExecute } from '../../src/adapter'
-import { TInputParameters } from '../../src/endpoint'
-
-describe('global endpoint', () => {
-  const jobID = '1'
-  const execute = makeExecute()
-  process.env.API_KEY = process.env.API_KEY ?? 'test_api_key'
-
-  describe('validation error', () => {
-    const requests = [
-      { name: 'empty body', testData: {} },
-      { name: 'empty data', testData: { data: {} } },
-      {
-        name: 'market not supplied',
-        testData: {
-          id: jobID,
-          data: { endpoint: 'globalMarketCap' },
-        },
-      },
-    ]
-
-    requests.forEach((req) => {
-      it(`${req.name}`, async () => {
-        try {
-          await execute(req.testData as AdapterRequest<TInputParameters>, {})
-        } catch (error) {
-          const errorResp = Requester.errored(jobID, error as AdapterError)
-          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
-        }
-      })
-    })
-  })
-})
