+++ packages/tsconfig.json
@@ -237,9 +237,6 @@
       "path": "./sources/coingecko"
     },
     {
-      "path": "./sources/coingecko-test"
-    },
-    {
       "path": "./sources/coinlore"
     },
     {
